Moved the notifications tab to the layout folder too.
