+++ src/app/components/layout/notifications/notifications.component.less
